[rename] phase.h -> common.h
